refactor: share update notification
